Misc. code cleanup, Preliminary IMGUI support

+++ Engine/FrameTimer.cpp
@@ -19,7 +19,7 @@ float FrameTimer::Mark()
 	duration<float> frameDuration = lastTime - old;
 	deltaTime = frameDuration.count();
 	auto totalTime = lastTime - startTime;
-	float total = totalTime.count() / 1000000000;
+	float total = (float) totalTime.count() / 1000000000;
 	currentFPS = 1 / deltaTime;
 	averageFPS = frameCount / total;
